Migrate code form app.rb to store_files.rb
