+++ sql/data.sql
@@ -1,0 +1,13 @@
+/* Populate database with sample data. */
+
+INSERT INTO genre (ID, NAME)
+VALUES (0, 'Rock');
+
+INSERT INTO genre (ID, NAME)
+VALUES (1, 'Pop');
+
+INSERT INTO genre (ID, NAME)
+VALUES (2, 'Comedy');
+
+INSERT INTO genre (ID, NAME)
+VALUES (3, 'Thriller');
